Extract HeatmapGridMetrics and intensityColor into a shared file
Pulls the layout-metric struct, GeometryReader preference key, and the
faint-blue-to-warm-orange gradient out of UsageHeatmapView and into a
new UsageHeatmapShared.swift so the upcoming UsageActivityView can
reuse them. UsageHeatmapView's file-private duplicates are deleted in
the same commit because Swift treats same-named file-private + module-
internal symbols as ambiguous for type lookup. Old view keeps its
private intensityColor method (no collision with the new top-level
free function).
